requirements.txt
scikit-learn>=1.3.2
xgboost>=1.7.6
pandas>=2.2.3
numpy>=1.26.5
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
shap>=0.44.0
lime>=0.2.0
joblib>=1.3.2
pytest>=7.4.2
snowflake-connector-python>=3.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.2
python-multipart>=0.0.6
httpx>=0.25.2  # For async HTTP
prometheus-client>=0.19.0  # For metrics
PyYAML>=6.0.1  # For config files
dotenv>=1.0.0  # For environment variable management